WIP: move annotation form into ChangeView; remove Version from PageDetails
