+++ src/components/page-details.tsx
@@ -1,8 +1,7 @@
 import * as PropTypes from 'prop-types';
 import * as React from 'react';
 import {Link, RouteComponentProps} from 'react-router-dom';
-import WebMonitoringDb, {Annotation, Page, Version} from '../services/web-monitoring-db';
-import AnnotationForm from './annotation-form';
+import WebMonitoringDb, {Page} from '../services/web-monitoring-db';
 import ChangeView from './change-view';
 
 // export type IPageDetailsProps = RouteComponentProps<{pageId: string}>;
@@ -13,9 +12,6 @@ export interface IPageDetailsProps extends RouteComponentProps<{pageId: string}>
 
 interface IPageDetailsState {
     page: Page;
-    version: Version;
-    annotation: any;
-    collapsedView: boolean;
 }
 
 export default class PageDetails extends React.Component<IPageDetailsProps, IPageDetailsState> {
@@ -27,10 +23,7 @@ export default class PageDetails extends React.Component<IPageDetailsProps, IPag
 
     constructor (props: IPageDetailsProps) {
         super(props);
-        this.state = {annotation: null, page: null, version: null, collapsedView: true};
-        this.saveAnnotation = this.saveAnnotation.bind(this);
-        this.updateAnnotation = this.updateAnnotation.bind(this);
-        this.toggleCollapsedView = this.toggleCollapsedView.bind(this);
+        this.state = {page: null};
     }
 
     componentWillMount () {
@@ -65,60 +58,25 @@ export default class PageDetails extends React.Component<IPageDetailsProps, IPag
         };
 
         // TODO: should factor out a loading view
-        if (!(this.state.page && this.state.version)) {
+        if (!(this.state.page)) {
             return <div>Loading…</div>;
         }
 
         const page = this.state.page;
-        const version = this.state.version;
-        const markAsSignificant = () => console.error('markAsSignificant not implemented');
-        const addToDictionary = () => console.error('markAsSignificant not implemented');
-
-        const diffLinks = [];
-        const versionista = version.source_metadata;
-        if (versionista.diff_with_previous_url) {
-            diffLinks.push(
-                <a href={versionista.diff_with_previous_url} target="_blank" rel="noopener">
-                    Last two diff
-                </a>
-            );
-        }
-        if (versionista.diff_with_first_url) {
-            diffLinks.push(
-                <a href={versionista.diff_with_first_url} target="_blank" rel="noopener">
-                    Last to base diff
-                </a>
-            );
-        }
 
         // TODO: this HTML should probably be broken up a bit
         return (
             <div className="container-fluid container-page-view">
                 <div className="row">
                     <div className="col-md-9">
-                        <h2 className="diff-title">{version.uuid} - {page.title}</h2>
+                        <h2 className="diff-title">{page.title}</h2>
                         <a className="diff_page_url" href={page.url} target="_blank" rel="noopener">{page.url}</a>
                     </div>
                     <div className="col-md-3">
                         {this.renderPager()}
                     </div>
                 </div>
-
-                <hr />
-
-                {this.renderAnalysisTools(markAsSignificant, addToDictionary)}
-
-                <hr />
-                <div className="row">
-                    <div className="col-md-12">
-                        {diffLinks[0]} || {diffLinks[1]}
-                    </div>
-                </div>
-                <div className="row">
-                    <div className="col-md-12">
-                        <ChangeView page={this.state.page} />
-                    </div>
-                </div>
+            <ChangeView page={this.state.page} />
             </div>
         );
     }
@@ -149,40 +107,6 @@ export default class PageDetails extends React.Component<IPageDetailsProps, IPag
         );
     }
 
-    private renderAnalysisTools (markAsSignificant: any, addToDictionary: any) {
-        if (!this.props.user) {
-            return <p>Log in to annotate changes.</p>;
-        }
-
-        return (
-            <div>
-                <div className="row">
-                    <div className="col-md-6">
-                        <i className="fa fa-toggle-on" aria-hidden="true" />
-                        {/* TODO: should be buttons */}
-                        <a className="lnk-action" href="#" onClick={this.toggleCollapsedView}>Toggle Signifiers</a>
-                        <i className="fa fa-pencil" aria-hidden="true" />
-                        <a className="lnk-action" href="#" onClick={this.saveAnnotation}>Update Record</a>
-                        <i className="fa fa-list" aria-hidden="true" />
-                        <Link to="/" className="lnk-action">Back to list view</Link>
-                    </div>
-                    <div className="col-md-6 text-right">
-                        <i className="fa fa-upload" aria-hidden="true" />
-                        <a className="lnk-action" href="#" onClick={markAsSignificant}>Add Important Change</a>
-                        <i className="fa fa-database" aria-hidden="true" />
-                        <a href="#" onClick={addToDictionary}>Add to Dictionary</a>
-                    </div>
-                </div>
-
-                <AnnotationForm
-                    annotation={this.state.annotation}
-                    onChange={this.updateAnnotation}
-                    collapsed={this.state.collapsedView}
-                />
-            </div>
-        );
-    }
-
     private loadPage (pageId: string) {
         const fromList = this.props.pages && this.props.pages.find(
             (page: Page) => page.uuid === pageId);
@@ -190,61 +114,7 @@ export default class PageDetails extends React.Component<IPageDetailsProps, IPag
         Promise.resolve(fromList || this.context.api.getPage(pageId))
             .then((page: Page) => {
                 this.setState({page});
-                this.loadLatestVersion(page);
             });
     }
 
-    private loadLatestVersion (page: Page) {
-        const latest = page.latest || page.versions[0];
-        if (!latest) {
-            this.loadPage(page.uuid);
-        }
-        else {
-            this.context.api.getVersion(page.uuid, latest.uuid)
-                .then(version => {
-                    if (this.state.page.uuid === version.page_uuid) {
-                        this.setState({
-                            annotation: version.current_annotation,
-                            version
-                        });
-                    }
-                });
-        }
-    }
-
-    private loadVersion (page: Page, version: Version) {
-        // const latest = page.latest || page.versions[0];
-        // if (!latest) {
-            // this.loadPage(page.uuid);
-        // }
-        // else {
-            // this.context.api.getVersion(page.uuid, latest.uuid)
-            //     .then(version => {
-            //         if (this.state.page.uuid === version.page_uuid) {
-            //             this.setState({
-            //                 annotation: version.current_annotation,
-            //                 version
-            //             });
-            //         }
-            //     });
-        // }
-    }
-
-    private updateAnnotation (newAnnotation: any) {
-        this.setState({annotation: newAnnotation});
-    }
-
-    private saveAnnotation (event: React.SyntheticEvent<HTMLElement>) {
-        event.preventDefault();
-        // TODO: display some indicator that saving is happening/complete
-        const version = this.state.version;
-        this.context.api.annotateVersion(version.page_uuid, version.uuid, this.state.annotation)
-            .then((annotationRecord: Annotation) => {
-                this.loadLatestVersion(this.state.page);
-            });
-    }
-
-    private toggleCollapsedView () {
-        this.setState(previousState => ({collapsedView: !previousState.collapsedView}));
-    }
 }
